:bulb: Implement simple caching with Redis (#27)
This will be one of my first opinionated stances on caching
implementations. As Redis is open source and fairly prevalent, the
configuration values for external cache shall be apart of `AppOptions`.

Furthermore, this PR enhances two GitHub calls by wrapping them in a
cache call. This will help reduce the amount of calls made to GitHub,
increasing performance overall.

Implementation wise, this PR does some bad things to the cleanliness of
this code (circular dependencies between modules, etc.). I am saying
that this is "okay" for now as an upcoming change will refactor this
repo heavily so as to leverage a DI container.